test: update filenames and class names for unit tests
