+++ test/t/unit/Makefile.am
@@ -4,7 +4,17 @@ EXTRA_DIST = \
 	test_unit_compgen.py \
 	test_unit_compgen_available_interfaces.py \
 	test_unit_compgen_commands.py \
+	test_unit_compgen_filedir.py \
+	test_unit_compgen_ip_addresses.py \
+	test_unit_compgen_known_hosts.py \
+	test_unit_compgen_pgids.py \
+	test_unit_compgen_pids.py \
+	test_unit_compgen_pnames.py \
 	test_unit_compgen_split.py \
+	test_unit_compgen_tilde.py \
+	test_unit_compgen_variables.py \
+	test_unit_compgen_xinetd_services.py \
+	test_unit_complete_longopt.py \
 	test_unit_count_args.py \
 	test_unit_delimited.py \
 	test_unit_deprecate_func.py \
@@ -12,31 +22,21 @@ EXTRA_DIST = \
 	test_unit_expand.py \
 	test_unit_expand_glob.py \
 	test_unit_expand_tilde.py \
-	test_unit_filedir.py \
 	test_unit_find_unique_completion_pair.py \
-	test_unit_get_first_arg.py \
 	test_unit_get_cword.py \
+	test_unit_get_first_arg.py \
 	test_unit_get_words.py \
 	test_unit_initialize.py \
-	test_unit_ip_addresses.py \
-	test_unit_known_hosts.py \
 	test_unit_load.py \
-	test_unit_longopt.py \
 	test_unit_looks_like_path.py \
 	test_unit_parse_help.py \
 	test_unit_parse_usage.py \
-	test_unit_pgids.py \
-	test_unit_pids.py \
-	test_unit_pnames.py \
 	test_unit_quote.py \
 	test_unit_quote_compgen.py \
 	test_unit_realcommand.py \
 	test_unit_split.py \
-	test_unit_tilde.py \
 	test_unit_unlocal.py \
-	test_unit_variables.py \
-	test_unit_xfunc.py \
-	test_unit_xinetd_services.py
+	test_unit_xfunc.py
 
 all:
