+++ test/t/unit/test_unit_compgen_tilde.py
@@ -4,7 +4,7 @@ from conftest import assert_bash_exec
 
 
 @pytest.mark.bashcomp(cmd=None, ignore_env=r"^\+COMPREPLY=")
-class TestUnitTilde:
+class TestUnitCompgenTilde:
     def test_1(self, bash):
         assert_bash_exec(bash, "! _comp_compgen_tilde >/dev/null")
