+++ test/t/unit/test_unit_compgen_xinetd_services.py
@@ -4,7 +4,7 @@ from conftest import assert_bash_exec
 
 
 @pytest.mark.bashcomp(cmd=None, ignore_env=r"^\+COMPREPLY=")
-class TestUnitXinetdServices:
+class TestUnitCompgenXinetdServices:
     def test_direct(self, bash):
         assert_bash_exec(bash, "_comp_compgen_xinetd_services >/dev/null")
